Fix build issue
Error appears to be two spaces different on Semaphore to VM, which
would suggest it thinks a tab is 2 spaces instead of 4. This could be
due to the OS or the XML library perhaps. Either way, we only care that
the error is converted to an exception, we’ve no control over its
contents anyway.